move xcm config to seperate file
