+++ runtime/karura/src/integration_tests_config.rs
@@ -16,6 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+use super::xcm_config::*;
 use super::*;
 
 parameter_types! {
